+++ infra/core/dev/providers.tf
@@ -19,6 +19,11 @@ terraform {
       source  = "pagopa-dx/aws"
       version = "~> 0.0"
     }
+
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 3.8"
+    }
   }
 
   backend "azurerm" {