+++ infra/modules/azure_core_infra/_modules/github_runner/container_app_environment.tf
@@ -14,6 +14,18 @@ resource "azurerm_container_app_environment" "cae" {
   infrastructure_subnet_id       = var.runner_snet
   zone_redundancy_enabled        = false
   internal_load_balancer_enabled = false
+  public_network_access          = "Disabled"
+
+  workload_profile {
+    name                  = "Consumption"
+    workload_profile_type = "Consumption"
+  }
+
+  lifecycle {
+    ignore_changes = [
+      infrastructure_resource_group_name, # Otherwise Terraform forces the recreation at every plan due to provider issue
+    ]
+  }
 
   tags = var.tags
 }