+++ infra/modules/azure_federated_identity_with_github/id.tf
@@ -21,21 +21,19 @@ resource "azurerm_user_assigned_identity" "cd" {
 resource "azurerm_federated_identity_credential" "ci_github" {
   count = local.is_ci_enabled
 
-  resource_group_name = local.resource_group_name
-  name                = format(local.federation_prefix, var.identity_type, "ci")
-  audience            = local.ci_github_federations.audience
-  issuer              = local.ci_github_federations.issuer
-  parent_id           = azurerm_user_assigned_identity.ci[0].id
-  subject             = local.ci_github_federations.subject
+  name      = format(local.federation_prefix, var.identity_type, "ci")
+  audience  = local.ci_github_federations.audience
+  issuer    = local.ci_github_federations.issuer
+  parent_id = azurerm_user_assigned_identity.ci[0].id
+  subject   = local.ci_github_federations.subject
 }
 
 resource "azurerm_federated_identity_credential" "cd_github" {
   count = local.is_cd_enabled
 
-  resource_group_name = local.resource_group_name
-  name                = format(local.federation_prefix, var.identity_type, "cd")
-  audience            = local.cd_github_federations.audience
-  issuer              = local.cd_github_federations.issuer
-  parent_id           = azurerm_user_assigned_identity.cd[0].id
-  subject             = local.cd_github_federations.subject
+  name      = format(local.federation_prefix, var.identity_type, "cd")
+  audience  = local.cd_github_federations.audience
+  issuer    = local.cd_github_federations.issuer
+  parent_id = azurerm_user_assigned_identity.cd[0].id
+  subject   = local.cd_github_federations.subject
 }