+++ infra/modules/azure_github_environment_bootstrap/id_app.tf
@@ -7,12 +7,11 @@ resource "azurerm_user_assigned_identity" "app_ci" {
 }
 
 resource "azurerm_federated_identity_credential" "github_app_ci" {
-  resource_group_name = azurerm_user_assigned_identity.app_ci.resource_group_name
-  name                = format(local.ids.federated_identity_name, "app", "ci")
-  audience            = local.ids.audience
-  issuer              = local.ids.issuer
-  parent_id           = azurerm_user_assigned_identity.app_ci.id
-  subject             = "repo:pagopa/${var.repository.name}:environment:${format(local.ids.app_environment_name, "ci")}"
+  name      = format(local.ids.federated_identity_name, "app", "ci")
+  audience  = local.ids.audience
+  issuer    = local.ids.issuer
+  parent_id = azurerm_user_assigned_identity.app_ci.id
+  subject   = "repo:pagopa/${var.repository.name}:environment:${format(local.ids.app_environment_name, "ci")}"
 }
 
 
@@ -25,10 +24,9 @@ resource "azurerm_user_assigned_identity" "app_cd" {
 }
 
 resource "azurerm_federated_identity_credential" "github_app_cd" {
-  resource_group_name = azurerm_user_assigned_identity.app_cd.resource_group_name
-  name                = format(local.ids.federated_identity_name, "app", "cd")
-  audience            = local.ids.audience
-  issuer              = local.ids.issuer
-  parent_id           = azurerm_user_assigned_identity.app_cd.id
-  subject             = "repo:pagopa/${var.repository.name}:environment:${format(local.ids.app_environment_name, "cd")}"
+  name      = format(local.ids.federated_identity_name, "app", "cd")
+  audience  = local.ids.audience
+  issuer    = local.ids.issuer
+  parent_id = azurerm_user_assigned_identity.app_cd.id
+  subject   = "repo:pagopa/${var.repository.name}:environment:${format(local.ids.app_environment_name, "cd")}"
 }